Cleaning up webcam nonsense from main and gado_sys

+++ main.py
@@ -20,24 +20,8 @@ from gado.gado_sys import GadoSystem
 from gado.db import DBFactory, DBInterface
 from Tkinter import Tk
 import threading
-from gado.Webcam import *
 
-        
-if __name__ == '__main__':
-    print "Initializing Gado Robot Management Interface"
-    
-    #Test picture taking
-    camera = Webcam()
-    print "Connected to webcam: %s" % str(camera.connect())
-    #camera.saveImage("superTest.jpg", camera.returnImage())
-    
-    # Import current gado settings
-    settings = import_settings()
-    
-    # Get access to the DB
-    db = DBFactory(**settings).get_db()
-    
-    
+def create_dummy_artifact_sets(db, clear=True):
     '''
     a
       b
@@ -49,9 +33,11 @@ if __name__ == '__main__':
       h
         i
     '''
-    db(db.artifacts.id > 0).delete()
-    db(db.artifact_sets.id > 0).delete()
     
+    if clear:
+        db(db.artifacts.id > 0).delete()
+        db(db.artifact_sets.id > 0).delete()
+        
     a_id = db.artifact_sets.insert(name='a', parent=None)
     b_id = db.artifact_sets.insert(name='b', parent=a_id)
     c_id = db.artifact_sets.insert(name='c', parent=b_id)
@@ -66,9 +52,18 @@ if __name__ == '__main__':
     db.artifacts.insert(artifact_set=d_id)
     db.artifacts.insert(artifact_set=i_id)
     db.artifacts.insert(artifact_set=f_id)
+        
+if __name__ == '__main__':
+    print "Initializing Gado Robot Management Interface"
     
+    # Import current gado settings
+    settings = import_settings()
     
+    # Get access to the DB
+    db = DBFactory(**settings).get_db()
     db_interface = DBInterface(db)
+    
+    create_dummy_artifact_sets(db)
     
     #Create instance of robot
     gado = Robot(**settings)
@@ -77,7 +72,7 @@ if __name__ == '__main__':
     tk = Tk()
     
     #Start up the Gado System for management
-    gado_sys = GadoSystem(db_interface, gado, camera, tk)
+    gado_sys = GadoSystem(db_interface, gado, tk)
     
     #Create the window which will display the autoconnect progress bar
     progressBar = ProgressBar(root=tk)
